third_party/blink/web_tests/platform/linux/virtual/url-non-special/external/wpt/url/url-setters.any_include=javascript-expected.txt

This is a testharness.js-based test.
[FAIL] URL: Setting <javascript://x/>.port = '12'
  assert_equals: expected "javascript://x:12/" but got "javascript://x/"
Harness: the test ran to completion.
